pyproject.toml
[tool.poetry]
name = "indexmaker"
version = "0.1.0"
description = ""
authors = ["nicolas <nsocri@ibmb.csic.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pdfplumber = "^0.11.7"
python-docx = "^1.2.0"
nltk = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
